zbuild: move parser to separate file
